Restructure - Wide to Long (R Script)
The script first makes a restructured data frame of the fixed and second dataframe in the list. Then it loops through the other frames, restructuring and adding them to the first restructure. Then it adds a numerical index of variable level at the beginning of the final restructured frame, saves it as "Restructure Data.csv" and calls a view of the final dataframe.